# docs/commons/openapi.yaml
openapi: 3.0.3
info:
  title: Reals Backend API
  version: 1.0.0
  description: Current backend contract for the Reals mobile client.
servers:
  - url: http://10.0.2.2:8080
    description: Android emulator against local backend
  - url: http://localhost:8080
    description: Local backend from the same machine
security:
  - firebaseAppCheck: []
    bearerAuth: []
...
paths:
...
  /api/chats/{chatId}:
    get:
      tags: [Chats]
      summary: Get chat
      parameters:
        - $ref: '#/components/parameters/ChatId'
      responses:
        '200':
          $ref: '#/components/responses/Chat'
        '404':
          $ref: '#/components/responses/ChatNotFound'
  /api/chats/{chatId}/messages:
    post:
      tags: [Chats]
      summary: Send chat message
      description: Sends a message as the authenticated chat participant. First chats preserve existing `MUTUAL_CANCEL` pending-request behavior. For second chats, sends cancel pending mutual-completion or partner-inactivity requests before expiry; at or after partner-inactivity expiry the lifecycle transition commits and the message is rejected without insertion.
      parameters:
        - $ref: '#/components/parameters/ChatId'
      requestBody:
        $ref: '#/components/requestBodies/SendMessage'
      responses:
        '200':
          description: Created message
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatMessageResponse'
        '400':
          $ref: '#/components/responses/ChatMessageInvalid'
        '404':
          $ref: '#/components/responses/ChatNotFound'
        '409':
          $ref: '#/components/responses/ChatAvailabilityConflict'
    get:
      tags: [Chats]
      summary: List chat messages
      description: Without a cursor, returns the legacy JSON array containing the most recent bounded page in chronological order. With `after` or `afterMessageId`, returns `ChatMessagesResponse` with messages strictly after the cursor ordered by `sentAt ASC, id ASC`; `hasMore` is true when additional rows exist beyond the returned page. Response formats are preserved for compatibility.
      parameters:
        - $ref: '#/components/parameters/ChatId'
        - name: after
          in: query
          required: false
          schema: { type: string, format: uuid }
          description: Return messages after this message id using the incremental response shape.
        - name: afterMessageId
          in: query
          required: false
          schema: { type: string, format: uuid }
          description: Alias for `after`.
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 200
          description: Maximum number of messages returned. Initial legacy reads fetch the most recent `limit` messages and return them chronologically. Incremental reads fetch at most `limit` messages after the cursor and use `limit + 1` internally to compute `hasMore`.
      responses:
        '200':
          description: Messages. Without a cursor this is the legacy array; with a cursor this is ChatMessagesResponse.
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/ChatMessageResponse'
                  - $ref: '#/components/schemas/ChatMessagesResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/ChatNotFound'
        '409':
          $ref: '#/components/responses/ChatAvailabilityConflict'
  /api/chats/{chatId}/audio-messages:
    post:
      tags: [Chats]
      summary: Send chat audio message
      description: Sends an audio message into the existing ordered chat stream. Requires multipart `file` and client-generated UUID `clientMessageId`. New messages return 201; idempotent replay with matching SHA-256 returns 200; same id with different content returns CHAT_MESSAGE_IDEMPOTENCY_CONFLICT.
      parameters:
        - $ref: '#/components/parameters/ChatId'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, clientMessageId]
              properties:
                file:
                  type: string
                  format: binary
                  description: MPEG-4/M4A audio-only AAC file, max 2 MiB and max 60 seconds.
                clientMessageId:
                  type: string
                  format: uuid
      responses:
        '200':
          description: Existing message returned for an idempotent replay
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatMessageResponse'
        '201':
          description: Created audio message
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatMessageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/ChatNotFound'
        '409':
          $ref: '#/components/responses/ChatAvailabilityConflict'
        '503':
          description: Audio upload pipeline is busy
          headers:
            Retry-After:
              schema: { type: integer }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/chats/{chatId}/guidance/next-request:
    post:
      tags: [Chats]
      summary: Request next first-chat guidance question
      description: Authenticated first-chat participant requests advancement from the current guided question. The request has no body. The question advances only after both participants have independently requested it and each requester has met the configured participation threshold for the active question interval. Partner request/readiness metadata is never returned. While a `MUTUAL_CANCEL` exit request is `PENDING`, guidance next requests are rejected with `CHAT_MUTUAL_CANCELLATION_PENDING` and guidance timestamps are not changed.
      parameters:
        - $ref: '#/components/parameters/ChatId'
      responses:
        '200':
          description: Current user-scoped first-chat guidance state after applying the request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FirstChatGuidanceResponse'
        '404':
          $ref: '#/components/responses/ChatNotFound'
        '409':
          $ref: '#/components/responses/FirstChatGuidanceConflict'
  /api/chats/{chatId}/exit-requests:
    post:
      tags: [Chats]
      summary: Request mutual chat cancellation
      description: Idempotent for the same requester while their mutual cancellation request is pending. A new request returns 201 Created. A repeated request from the same requester returns 200 OK with the existing request and does not overwrite reason or details. A pending mutual request created by the partner remains a conflict. A pending mutual cancellation pauses new message sends and first-chat guidance advancement, while message reads, exit-request polling, acceptance, rejection, timeout, unilateral cancellation and safety cancellation remain available.
      parameters:
        - $ref: '#/components/parameters/ChatId'
      requestBody:
        $ref: '#/components/requestBodies/ChatExitRequestCreate'
      responses:
        '200':
          description: Existing pending exit request for the same requester
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatExitRequestResponse'
        '201':
          description: Created exit request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatExitRequestResponse'
        '400':
          $ref: '#/components/responses/ChatMessageInvalid'
        '404':
          $ref: '#/components/responses/ChatNotFound'
        '409':
          $ref: '#/components/responses/ChatExitRequestCreateConflict'
    get:
      tags: [Chats]
      summary: List visible exit requests for chat participant
      parameters:
        - $ref: '#/components/parameters/ChatId'
      responses:
        '200':
          description: Exit requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChatExitRequestResponse'
        '404':
          $ref: '#/components/responses/ChatNotFound'
  /api/chats/{chatId}/exit-requests/{exitRequestId}/acceptance:
    post:
      tags: [Chats]
      summary: Accept mutual cancellation request
      description: Accepts a pending mutual cancellation request and closes the chat without penalty.
      parameters:
        - $ref: '#/components/parameters/ChatId'
        - $ref: '#/components/parameters/ExitRequestId'
      responses:
        '200':
          $ref: '#/components/responses/ChatExitOutcome'
        '404':
          $ref: '#/components/responses/ChatExitRequestNotFound'
        '409':
          $ref: '#/components/responses/ChatExitRequestActionConflict'
  /api/chats/{chatId}/exit-requests/{exitRequestId}/rejection:
    post:
      tags: [Chats]
      summary: Reject mutual cancellation request
      description: Rejects a pending mutual cancellation request and closes the chat. Future scoring may apply a lower penalty to the requester, but no penalty is applied today.
      parameters:
        - $ref: '#/components/parameters/ChatId'
        - $ref: '#/components/parameters/ExitRequestId'
      responses:
        '200':
          $ref: '#/components/responses/ChatExitOutcome'
        '404':
          $ref: '#/components/responses/ChatExitRequestNotFound'
        '409':
          $ref: '#/components/responses/ChatExitRequestActionConflict'
  /api/chats/{chatId}/exit-requests/{exitRequestId}/timeout:
    post:
      tags: [Chats]
      summary: Timeout mutual cancellation request
      description: Resolves an unanswered mutual cancellation request after the configured timeout and closes the chat without penalty. This is not treated as a unilateral cancellation; if the requester calls it because the responder did not answer in time, the requester must not be penalized.
      parameters:
        - $ref: '#/components/parameters/ChatId'
        - $ref: '#/components/parameters/ExitRequestId'
      responses:
        '200':
          $ref: '#/components/responses/ChatExitOutcome'
        '404':
          $ref: '#/components/responses/ChatExitRequestNotFound'
        '409':
          $ref: '#/components/responses/ChatExitRequestActionConflict'
  /api/chats/{chatId}/cancellations:
    post:
      tags: [Chats]
      summary: Unilateral chat cancellation
      parameters:
        - $ref: '#/components/parameters/ChatId'
      requestBody:
        $ref: '#/components/requestBodies/ChatCancellation'
      responses:
        '201':
          $ref: '#/components/responses/ChatExitOutcome'
        '400':
          $ref: '#/components/responses/ChatMessageInvalid'
        '404':
          $ref: '#/components/responses/ChatNotFound'
        '409':
          $ref: '#/components/responses/ChatAvailabilityConflict'
  /api/chats/{chatId}/safety-cancellations:
    post:
      tags: [Chats]
      summary: Safety/report chat cancellation
      description: Closes the chat and creates a pending internal safety report. Reporting does not automatically penalize the reported participant; the response keeps penaltyApplied=false until an admin confirms the report separately.
      parameters:
        - $ref: '#/components/parameters/ChatId'
      requestBody:
        $ref: '#/components/requestBodies/ChatSafetyCancellation'
      responses:
        '201':
          $ref: '#/components/responses/ChatExitOutcome'
        '400':
          $ref: '#/components/responses/ChatMessageInvalid'
        '404':
          $ref: '#/components/responses/ChatNotFound'
        '409':
          $ref: '#/components/responses/ChatAvailabilityConflict'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Firebase ID token
    firebaseAppCheck:
      type: apiKey
      in: header
      name: X-Firebase-AppCheck
      description: Firebase App Check token for Android-facing API requests.
  parameters:
...
    ChatId:
      name: chatId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    ExitRequestId:
      name: exitRequestId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  requestBodies:
...
    SendMessage:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SendMessageRequest'
    ChatExitRequestCreate:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ChatExitRequestCreateRequest'
    ChatCancellation:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ChatCancellationRequest'
    ChatSafetyCancellation:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ChatSafetyCancellationRequest'
...
  responses:
...
    Chat:
      description: Chat
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ChatResponse'
    ChatExitOutcome:
      description: Chat exit outcome
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ChatExitOutcomeResponse'
...
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ChatMessageInvalid:
      description: "Invalid chat message or chat-action text. Stable code: `CHAT_MESSAGE_INVALID`."
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: CHAT_MESSAGE_INVALID
            error: Bad Request
            message: Chat message is invalid
    ChatNotFound:
      description: "Chat not found. Stable code: `CHAT_NOT_FOUND`."
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: CHAT_NOT_FOUND
            error: Not Found
            message: Chat was not found
    ChatAvailabilityConflict:
      description: Chat is not currently readable or writable. Message sends may also return `CHAT_MUTUAL_CANCELLATION_PENDING`; active chat participation writes may also return `LEGAL_ACTION_REQUIRED`.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            legalActionRequired:
              value:
                code: LEGAL_ACTION_REQUIRED
                error: Conflict
                message: Current legal document actions are required before continuing
            chatNotAvailable:
              value:
                code: CHAT_NOT_AVAILABLE
                error: Conflict
                message: Chat is not available
            chatExpired:
              value:
                code: CHAT_EXPIRED
                error: Conflict
                message: Chat has expired
            mutualCancellationPending:
              value:
                code: CHAT_MUTUAL_CANCELLATION_PENDING
                error: Conflict
                message: A mutual cancellation request is pending
    FirstChatGuidanceConflict:
      description: First-chat guidance cannot be advanced in the current state. Requesting the next guidance question may also return `CHAT_MUTUAL_CANCELLATION_PENDING` or `LEGAL_ACTION_REQUIRED`.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            legalActionRequired:
              value:
                code: LEGAL_ACTION_REQUIRED
                error: Conflict
                message: Current legal document actions are required before continuing
            participationRequired:
              value:
                code: FIRST_CHAT_GUIDANCE_PARTICIPATION_REQUIRED
                error: Conflict
                message: Minimum participation is required before requesting another question
            alreadyRequested:
              value:
                code: FIRST_CHAT_GUIDANCE_NEXT_ALREADY_REQUESTED
                error: Conflict
                message: Next question has already been requested
            completed:
              value:
                code: FIRST_CHAT_GUIDANCE_COMPLETED
                error: Conflict
                message: First-chat guidance is completed
            chatNotAvailable:
              value:
                code: CHAT_NOT_AVAILABLE
                error: Conflict
                message: Chat is not available
            mutualCancellationPending:
              value:
                code: CHAT_MUTUAL_CANCELLATION_PENDING
                error: Conflict
                message: A mutual cancellation request is pending
...
    ChatExitRequestNotFound:
      description: Chat or exit request not found. Missing exit requests use stable code `CHAT_EXIT_REQUEST_NOT_FOUND`.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            chatNotFound:
              value:
                code: CHAT_NOT_FOUND
                error: Not Found
                message: Chat was not found
            exitRequestNotFound:
              value:
                code: CHAT_EXIT_REQUEST_NOT_FOUND
                error: Not Found
                message: Chat exit request was not found
    ChatExitRequestCreateConflict:
      description: Mutual chat cancellation request cannot be created in the current chat state.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            chatNotAvailable:
              value:
                code: CHAT_NOT_AVAILABLE
                error: Conflict
                message: Chat is not available
            chatExpired:
              value:
                code: CHAT_EXPIRED
                error: Conflict
                message: Chat has expired
            exitRequestAlreadyPending:
              value:
                code: CHAT_EXIT_REQUEST_ALREADY_PENDING
                error: Conflict
                message: A chat exit request is already pending
    ChatExitRequestActionConflict:
      description: Chat exit request cannot be accepted, rejected, or timed out in the current state.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            chatNotAvailable:
              value:
                code: CHAT_NOT_AVAILABLE
                error: Conflict
                message: Chat is not available
            chatExpired:
              value:
                code: CHAT_EXPIRED
                error: Conflict
                message: Chat has expired
            exitRequestNotAvailable:
              value:
                code: CHAT_EXIT_REQUEST_NOT_AVAILABLE
                error: Conflict
                message: Chat exit request is not available
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [code, error, message]
      properties:
        code: { type: string, example: PROFILE_NOT_ACTIVE }
        error: { type: string, example: Conflict }
        message: { type: string, nullable: true }
...
    ChatResponse:
      type: object
      required: [id, matchId, chatType, status, startedAt, timeoutAt, expiresAt]
      properties:
        id: { type: string, format: uuid }
        matchId: { type: string, format: uuid }
        connectionId: { type: string, format: uuid, nullable: true }
        chatType: { $ref: '#/components/schemas/ChatType' }
        status: { $ref: '#/components/schemas/ChatStatus' }
        startedAt: { type: string, format: date-time }
        availableAt: { type: string, format: date-time, nullable: true }
        activatedAt: { type: string, format: date-time, nullable: true }
        conversationStartedAt: { type: string, format: date-time, nullable: true }
        timeoutAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time }
        endedAt: { type: string, format: date-time, nullable: true }
        readOnlyUntil: { type: string, format: date-time, nullable: true }
        lastMessageAt: { type: string, format: date-time, nullable: true }
        lastMessageSenderId: { type: string, format: uuid, nullable: true }
        audioPolicy:
          allOf:
            - $ref: '#/components/schemas/ChatAudioPolicyResponse'
          nullable: true
        inactivityExpiresAt:
          type: string
          format: date-time
          nullable: true
          description: First-chat inactivity deadline. Null for chat types where inactivity timeout does not apply.
        audioPolicy:
          allOf:
            - $ref: '#/components/schemas/ChatAudioPolicyResponse'
          nullable: true
...
    FirstChatGuidanceQuestionResponse:
      type: object
      required: [id, text]
      properties:
        id: { type: string }
        text: { type: string }
    FirstChatGuidanceResponse:
      type: object
      required:
        - question
        - questionOrdinal
        - maxQuestions
        - requiredCharacters
        - canRequestNext
        - myNextRequested
        - completed
      properties:
        question:
          $ref: '#/components/schemas/FirstChatGuidanceQuestionResponse'
        questionOrdinal:
          type: integer
          minimum: 1
          maximum: 3
        maxQuestions:
          type: integer
          example: 3
        requiredCharacters:
          type: integer
          example: 40
        canRequestNext:
          type: boolean
          description: Eligibility for the authenticated user only.
        myNextRequested:
          type: boolean
          description: Whether the authenticated user has requested continuation for the current question.
        completed:
          type: boolean
          description: True when the final configured guidance question is active. The final question remains returned as `question`, and no further next-request action is available.
    SendMessageRequest:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 1000
          pattern: '^[^\u0000-\u0009\u000B\u000C\u000E-\u001F\u007F-\u009F<>]*$'
          description: Plain-text chat content. Line breaks are allowed and normalized to LF by the backend.
    ChatMessageResponse:
      type: object
      required: [id, chatSessionId, senderId, messageType, sentAt]
      properties:
        id: { type: string, format: uuid }
        chatSessionId: { type: string, format: uuid }
        senderId: { type: string, format: uuid }
        clientMessageId: { type: string, format: uuid, nullable: true }
        messageType: { $ref: '#/components/schemas/ChatMessageType' }
        content: { type: string, nullable: true }
        audio:
          allOf:
            - $ref: '#/components/schemas/ChatMessageAudioResponse'
          nullable: true
        sentAt: { type: string, format: date-time }
...
    ChatMessagesResponse:
      type: object
      required: [messages, hasMore, serverTime]
      properties:
        messages:
          type: array
          items:
            $ref: '#/components/schemas/ChatMessageResponse'
        hasMore:
          type: boolean
          description: True when an incremental request has additional messages beyond the returned page.
        serverTime: { type: string, format: date-time }
...
    ChatExitRequestResponse:
      type: object
      required: [id, chatId, requesterUserId, responderUserId, type, status, createdAt]
      properties:
        id: { type: string, format: uuid }
        chatId: { type: string, format: uuid }
        requesterUserId: { type: string, format: uuid }
        responderUserId: { type: string, format: uuid }
        type: { $ref: '#/components/schemas/ChatExitRequestType' }
        status: { $ref: '#/components/schemas/ChatExitRequestStatus' }
        reason:
          allOf:
            - $ref: '#/components/schemas/ChatExitReason'
          nullable: true
        details: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        resolvedAt: { type: string, format: date-time, nullable: true }
...
    ChatMessageType:
      type: string
      enum: [TEXT, AUDIO]
...
    ChatExitRequestType:
      type: string
      enum: [MUTUAL_CANCEL, UNILATERAL_CANCEL, SAFETY_REPORT]
    ChatExitRequestStatus:
      type: string
      enum: [PENDING, ACCEPTED, REJECTED, TIMED_OUT]
...